Towers now cost full credit/time when switching effects. Changing tower targeting behaviors no longer requires apply button to be hit. Fixed mana infusion RPC errors.

+++ Assets/scripts/Game/GameControl.js
@@ -38,7 +38,7 @@ function Update()
          if (Time.time >= roundEndTime)
             EndRound();
 
-         var newInfusionSize : float;
+         var newInfusionSize : int;
          if (Time.time >= nextAttackInfusionTime)
          {
             newInfusionSize = Mathf.Lerp(
